Split pixelcolor module
